wgm_peer: Add `--bind-dev` option
An upcoming design will need --bind-ip to be combined with the interface
name, as such add a new option --bind-dev which is mandatory to be
provided if --bind-ip is provided. However, --bind-dev can be provided
without providing --bind-ip.

This is required to handle the packet after going through the iptables
mangle table.

Signed-off-by: Ammar Faizi <ammarfaizi2@gnuweeb.org>